[UPDATE] rename folders for existing products list component
